feat!: refactored for new images system
