+++ server/src/domain/brands/dto/create-brand.dto.ts
@@ -13,7 +13,7 @@ import {
   ValidateNested
 } from 'class-validator';
 
-import { MetadataDto } from 'shared/metadata.dto';
+import { MetadataDto } from 'shared/dto/metadata.dto';
 
 export class CreateBrandDto {
   @ApiProperty({
