--- server/src/domain/brands/dto/update-icon-brand.dto.ts
@@ -1,3 +1,0 @@
-export class UpdateIconBrandDto {
-  readonly icon: string;
-}
