+++ server/src/domain/brands/schemas/brand.schema.ts
@@ -1,16 +1,9 @@
-import {
-  Prop,
-  Schema,
-  SchemaFactory
-} from '@nestjs/mongoose';
+import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { ApiProperty } from '@nestjs/swagger';
-import {
-  Document,
-  HydratedDocument,
-  Types
-} from 'mongoose';
+import { Document, HydratedDocument, Types } from 'mongoose';
 
-import MetadataProps from 'shared/metadata-props.schema';
+import { Image } from 'domain/images/schemas/image.schema';
+import { Metadata } from 'shared/schemas/metadata.schema';
 
 export type BrandDocument = HydratedDocument<Brand>;
 
@@ -40,19 +33,19 @@ class Brand extends Document {
   @Prop({ type: String, required: true })
   readonly title: string;
 
-  @ApiProperty({ example: 'public/brands/icon.svg' })
-  @Prop({ type: String, default: null })
-  readonly icon: string;
+  @ApiProperty({ type: Image })
+  @Prop({ type: Image, default: null })
+  readonly icon: Image;
 
   @ApiProperty({ example: 'Reparing Apple phones...' })
   @Prop({ type: String })
   readonly article: string;
 
   @ApiProperty({
-    type: MetadataProps
+    type: Metadata
   })
-  @Prop({ type: MetadataProps })
-  readonly metadata: MetadataProps;
+  @Prop({ type: Metadata })
+  readonly metadata: Metadata;
 }
 
 const BrandSchema = SchemaFactory.createForClass(Brand);
