+++ server/src/domain/contacts/schemas/contact.schema.ts
@@ -2,6 +2,8 @@ import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { ApiProperty } from '@nestjs/swagger';
 import { Document, HydratedDocument, Types } from 'mongoose';
 
+import { Image } from 'domain/images/schemas/image.schema';
+
 export type ContactDocument = HydratedDocument<Contact>;
 
 class Coords {
@@ -30,6 +32,10 @@ class Contact extends Document {
   @ApiProperty({ example: 'Саперно-Слобідська, 10' })
   @Prop({ type: String, required: true })
   readonly address: string;
+
+  @ApiProperty({ type: Image })
+  @Prop({ type: Image, default: null })
+  readonly image: Image;
 
   @ApiProperty({
     example: 'Вхід через супермаркет ВЕЛМАРТ'
