--- server/src/domain/gadgets/dto/update-image-gadget.dto.ts
@@ -1,5 +1,0 @@
-export class UpdateImageGadgetDto {
-  readonly image?: string;
-  readonly gallery?: string[];
-  readonly icon?: string;
-}
