+++ server/src/shared/schemas/metadata.schema.ts
@@ -1,7 +1,7 @@
 import { Prop } from '@nestjs/mongoose';
 import { ApiProperty } from '@nestjs/swagger';
 
-class MetadataProps {
+export class Metadata {
   @ApiProperty({
     example: 'Reliable Maintenance and Restoration'
   })
@@ -20,5 +20,3 @@ class MetadataProps {
   @Prop({ type: String, required: true })
   readonly keywords: string;
 }
-
-export default MetadataProps;
